`SharedMimeInfo`: Small fixes, including selecting GCC 8 as default compiler
Hilariously, GCC <8 cannot find `-lpthread` on ppc64le.  I'm not sure
why this is, (something screwed up in the linker scripts, I imagine) but
since this is a pure-C library, using GCC 8 doesn't change anything for
us, so I'm not going to try and backport the fix.  Someday we may need
to, but today is not that day.