[Git] For Windows, simply repackage git-for-windows
I couldn’t do a single script for all platforms because git-for-windows is a
tarbomb that would explode in `$WORKSPACE/srcdir`, and there would also be a
conflict between the 32- and the 64-bit directories, even though we only care
about the mingw* directories.  In addition, these Windows tarbombs are
self-contained (they already ship all dependencies), so for Windows we really
don’t want to have dependencies.  In the end, `build_tarballs` needs
platform-dependent sources and dependencies.