[AZP] Fix `init_jll_package()` race condition
Because we have multiple workers building at the same time, and because
we share a Julia package repository, we can occasionally hit race
conditions such as having 8 workers all try to initialize a JLL package
directory at the same time.  This should avoid that, and also avoid the
ENOTFASTFORWARD errors we get from git every now and then when we have a
dirty JLL package directory lying around.