[rr]: Build for `x86_64-linux-gnu`
If we can't target `musl` yet, at least get `rr` for `x86_64-linux-gnu`.

This requires a nice set of hack-n-slash patches to work around "old"
glibc paired with a "new" kernel.  We're just on the cusp of `rr`'s
support (it looks like they implicitly require `2.14`, so it wasn't too
bad to patch support in by simply assuming that we have a new enough
kernel (which is a safe assumption as the features needed were added in
the `2.6.3X` era)..