Update rr commit sha
The current version segfaults when recording julia on
recent versions of master. This is fixed on rr master,
but there hasn't been an rr release yet. This builds
from master. Version is set to 5.3.1 which doesn't
exist upstream, but since there have been a number
of changes since 5.3.0, calling this that didn't
seem right.