[PlatformSupport] Link `${sysroot}/usr/local` to entire `${prefix}`
We're currently linking only `lib/` and `lib64/` to the prefix, however CMake
find modules often find libraries in `${sysroot}/usr/local/lib` and then use
relative paths to executables or header files, that therefore can't be found.
With this change we link `${sysroot}/usr/local` to the entire prefix, which
should alleviate these issues, until the time we don't need to add this link at
all.