Bump rr
This is rr master plus https://github.com/mozilla/rr/pull/2515 and one additional commit
I'm about to PR upstream. There's some question of whether the 2515 PR can be merged
in its current form, but for the moment it's required for us. We want to turn it on on CI, so
let's build a version that has it. We can revisit what to do once all the patch upstreaming is done.