[libcgal-julia]: Minor update to v0.9.1
Instead of working about a method to map circulators/iterators directly
to julia (even though they're mostly pointers and dereferncing them
should be enough, the underlying types are unnecessarily complex), the
elements are collected in the C++-side into arrays.

As such, this patch includes a tiny change which involves copying the
beginning of the circulator instead of modifying it directly.