Add binutils patch to force deterministic output for `dlltool`
Without this, when `dlltool` outputs a `.a` file, it does not do so
deterministically, embedding timestamps and whatnot.  In an analogous
way as `ar` and `ranlib` turn on deterministic `bfd` mode when `-D` is
passed, we enable setting the same flags.  I don't have time right now
to make this a nice patch to upstream with command line arguments and
configure-time options, but if someone wants to do so, it should be
pretty easy.