[GCCBootstrap]: Change ARM shard to support armv6l and armv7l
* Upgrade `GCCBootstrap` to Julia 1.6+ syntax

* Change GCC bootstrapping process to use armv6l as the default
microarchitecture.  This builds an armv6l-compatible version of the
initial compiler support units, and allows us to build armv6 or armv7
targeting binaries.  It may cause a slight slowdown on armv7l, (due to
some of those compiler support libraries using only older, slower, armv6
instructions).  If it turns out to be too much of a problem, we may need
to build independent copies of the arm shards.