Fix build_tarballs.jl for Arb_jll (#1754)
This was using a bad Dependency on FLINT_jll (my fault) which later was
fixed in the package registry. Let's fix it here, too, to ensure it is
not forgotten during the next real update. The existing binaries can
be reused, but let's regenerate the JLL to get a correct Project.toml.

[skip build]