Fix `libicudata` issues on armv7l
armv7l has a slightly more persnickety linker than on other platforms,
and the weirdness that ICU does here, trying to create a "data-only"
shared library causes the linker to freak out.  It turns out most
distros patch away some of the weirdness, so we carry a patch as well.
Since it is highly unlikely any Julia package will ever want to load
this data directly, we'll also mark it as `dont_dlopen` for good
measure.