[FFTW] Update FFTW source and use version 3.3.9
I have changed the download source for FFTW to be fftw's official download link,
instead of the link that was being here (an unofficial one). FFTW recently
released version 3.3.9, which includes at least one function that is not present
in the current FFTW_jll artifact. This is likely due to the current download
link pointing to an old release candidate for fftw 3.3.9, instead of the
released version.

More details about this change, and why I think it is necessary, can be found in
JuliaMath/FFTW.jl#170.