Add `UserNSSandbox` (#2399)
As part of the effort to split BB up into smaller, more manageable
pieces, create a standalone JLL that contains the `sandbox` user
namespacing utility, and strip out all parts of its life as an `init`
replacement for running inside of `QEMU`.  If we ever resurrect that
part of its life, we'll do so with a dedicated tool.