[CompilerSupportLibraries] Don't eliminate `.so` symlinks
We want the symlinks such as `libatomic.so -> libatomic.so.1`.  This
symlink was being eliminated in the `rm` pass that was trying to drop
all linker scripts, but because the check (which was attempting to
determine whether a library is a proper `ELF` file or not) did not
dereference symlinks, it was rejecting all symlinks as non-`ELF` files.