update Dex to v2.28.1, patch for multiple sessions (#2820)
* update Dex to v2.28.1, patch for multiple sessions

This updates Dex to v2.28.1.

In addition, this adds a patch to allow Dex, based on a configuration switch, to issue multiple refresh tokesn for the same user.

In the current design Dex stores a refresh token per user and connector. When a refresh token is encashed or when a new offline token is requested, Dex scans the offline sessions table for an existing offline session, retrieves the refresh token id from the offline sessions table, issues a new id token based on the data stored against the refresh token id, updates the refresh token, and returns the renewed id token and refresh token. Thus, the old refresh token is no longer valid.

With this change, if the option `enableMultiRefreshTokens` is set to `true` in the config (default is false), Dex would issue a new refresh token when a new offline token is requested instead of updating the old one. So it is possible to have a single user to have multiple valid offline tokens (both id and refresh) at the same time. It would however continue the old behavior of updating the refresh token when it is encashed.

This also adds a configurable expiry time for unused refresh tokens, which is set at 30 days by default. Unused refresh tokens beyond this time are invalidated, and user is required to log in again.

To summarize, additional entries in the configuration file to enable this feature look like this:

```
expiry:
  refreshTokens: "720h"
enableMultiRefreshTokens: true
```