[HarfBuzz] Upgrade to v2.8.1, split ICU, build for experimental platforms (#3069)
Depending on ICU is problematic as it breaks ABI at every major versions, which
is reasonable, but they release major versions very frequently.  However we
don't really need the ICU stuff for anything, so we're going to split the
ICU-related stuff to a separate package, as most distributions do as well.