[libcgal-julia] Fix gcc version (julia 1.3)
Turns out the gcc version was related to the julia version itself.
Local CI showed me the shared libraries contained references to
GLIBCXX_3.4.26 that wasn't available in the libstdc++ shipped with julia
versions <1.6.  If that is the case, a rebuild is required using a
different gcc version.  To that end, I'm reverting the change and
retrying to build 1.3 and up with gcc v8.  If that fixes the issue with
1.3, I'll submit PRs for 1.4 and 1.5.

If I'm off by a mile (today hasn't been the most consistent day), I
apologize in advance, this PR can be ignored.