Xyce: Expand supported platforms
The AC_F77_LIBRARY_LDFLAGS flag is required when linking C++ and
Fortran *source files* into the same (shared) object. However,
here we're merely linking a Fortran library, so it is not required.
Unfortunately, while it in theory should do no harm, in practice
the AC_F77_LIBRARY_LDFLAGS macro is broken when targeting Windows
(which I will attempt to fix separately in upstream autoconf),
so remove it here.