Wizard recipe: nativefiledialog-v1.1.6 (#3743)
* New Recipe: nativefiledialog v1.1.6

* Update build_tarballs.jl

Ok, so this is the update with the instructions to generate the correspondent shared libraries for every platform. The dependencies for GTK3 are at least for linking, I think (but haven't proved otherwise) if they can be set aside for running.

* Update build_tarballs.jl

Finally, hope this works. :D

* Update build_tarballs.jl

Merge last recomendations from Mosè.

* [NativeFileDialog] Rename from lowercase

Co-authored-by: Mosè Giordano <mose@gnu.org>