[libunwind] Add a separate builder for v1.3.2 (#3859)
PR #3856 refactored the previous master into a builder specifically for
libunwind v1.5.0. To avoid issues with automated JLL registration, it
only did that version. This commit now adds a separate builder for
v1.3.2, the version of libunwind currently in use by Julia. Note that to
avoid said issues with registration, this commit does not refactor the
other builder to use a common setup; that can be done separately to
reduce duplication.