libffi: Stop messing with the install diretory (#3906)
The libffi configure script tries to be "smart" and detect whether
the target platform uses the lib/lib64 split for its libraries even
though nobody asked it too. This causes libffi to not be linkable
because it doesn't get installed in the lib directory. Now, in theory
pkg-config would take care of that, because it does have the correct
directory, but in practice pkg-config does not work, so we really do
need everything in one directory. The last question may be why didn't
this cause trouble before? Well, the answer to that one is that on
many platforms, we have a copy of libffi in the sysroot, which
builds picked up instead of the correct dependency. We don't on
powerpc64 though, causing issues in #3903.

Fix all that by just patching the problematic code out of the
configure script. This seems to be a fairly common patch for
build systems to apply. E.g. there's a version of it in [1].

[1] https://github.com/chef/omnibus-software/blob/main/config/patches/libffi/libffi-3.2.1-disable-multi-os-directory.patch