Vendor GraphViz config file (#3903)
* WIP: Vendor GraphViz config file

GraphViz uses a file called `config6` to specify which plugins
are available for loading. Generally, package managers are expected
to run `dot -c` upon installation which will enumerate the plugins
and detect whether they are functional. Without running `dot -c`,
GraphViz will refuse to load any plugins and be essentially
non-functional. In our case, we can't really run `dot -c` for two
reasons:

1. It would write into the artifacts directory, which is a big no-no.
2. It would potentially make errors non-consistent even with the same
   version of the GraphViz_jll, because the availability of features
   would depend on GraphViz's dynamic scan (which shouldn't change,
   but definitely could if external packages are updated).

So to workaround that, ship a generated config6. Unfortunately,
we do probably have to split these out by platform (or at least
by linux/mac/windows), since plugin availability varies between
platforms. I've included the linux one here, since that's the platform
I have easily available. If this direction seems good, I'll add the
other ones (or other people can push whatever they have access to).

* Update build_tarballs.jl

* Add darwin config6

* graphviz: Fix Linux pangocairo plugin

* libffi: Stop messing with the install diretory

The libffi configure script tries to be "smart" and detect whether
the target platform uses the lib/lib64 split for its libraries even
though nobody asked it too. This causes libffi to not be linkable
because it doesn't get installed in the lib directory. Now, in theory
pkg-config would take care of that, because it does have the correct
directory, but in practice pkg-config does not work, so we really do
need everything in one directory. The last question may be why didn't
this cause trouble before? Well, the answer to that one is that on
many platforms, we have a copy of libffi in the sysroot, which
builds picked up instead of the correct dependency. We don't on
powerpc64 though, causing issues in #3903.

Fix all that by just patching the problematic code out of the
configure script. This seems to be a fairly common patch for
build systems to apply. E.g. there's a version of it in [1].

[1] https://github.com/chef/omnibus-software/blob/main/config/patches/libffi/libffi-3.2.1-disable-multi-os-directory.patch

* Update graphviz, enable windows/freebsd

* Add config6 for mingw

* Fix windows export name issue

Co-authored-by: Mosè Giordano <giordano@users.noreply.github.com>