Update RadeonProRender version, build on macos aarch64 (#4708)
* Update RadeonProRender version, build on macos aarch64

This updates the version of RadeonProRender to v2.02.12, released in March 2022.  Additionally, it adds Mac M1 as a build target, given that the website claims it can run on M1 chips as well.

* Update version number and change julia compat

Since Mac M1 requires Julia v1.6 or greater, this seems like the appropriate thing to do.  Suggested in https://github.com/JuliaPackaging/Yggdrasil/pull/4001#issuecomment-984564572.

* Update build_tarballs.jl

Co-authored-by: Mosè Giordano <giordano@users.noreply.github.com>