Bump Kokkos Version, Add C++17 Support
I'm working on a project that I'd like to put up on Yggdrasil, but it requires Kokkos to be built with C++17. I thought that simultaneously bumping the version would be an efficient way of adding that, while letting people version control if something goes awry with the C++17 option.