New Recipe: FlangClassic
This adds a new recipe for "flang classic" (not to be confused
with LLVM's new flang which cannot yet generate code) from
https://github.com/flang-compiler/flang. Currently only
x86_64-linux-musl is supported because the build does complicated
bootstrap things to build its runtime library. The intent here
is to try to this in sanitizer builds of fortran-containing jlls,
so the musl version is sufficient (though we'll need to figure out
how to build the rtlib for each target).