Add a standalone LibStdCxx jll
Rather than trying to shoehorn the libstdcxx build into bootstrap as in #5308,
just add a separate jll for it and in a follow-up PR, I'll adjust CSL to pull
libstdcxx from here instead. In the medium term future, we may want to change
this again, but for now, I think this is the path of least resistance.