Sundials: Build for msan
There have been some mysterious Sundials memory errors lately,
so there's been a request for running sundials with msan. It'll
also be useful to have a non-Base jll for testing that we can
have the artifacts system properly load the msan variants of
everything, so this will hopefully be useful for that.