Add `flashrom`
This is technically not `v1.2`, but instead some random `master`
version, because the board I actually need to flash is too new to appear
in `v1.2` proper.  So we denote this by calling it `v1.2.900`, in hopes
that there will not be a large number of upstream releases on the `v1.2`
branch.