[LibTracyClient] Revise selected feature set. (#6307)
This is a relatively opinionated set of (non-)features that we'd
like to support initially. Later, we can expand the available
product set depending on the use cases that are required.

The main items that we might want in the near future are:
  - build w/ TRACY_ON_DEMAND disabled
  - build w/ TRACY_DELAYED_INIT enabled
       (and possibly TRACY_MANUAL_LIFETIME enabled, too)

Unfortunately, I don't see a way around that product matrix, since
Tracy is very low-overhead and it can be relatively costly to switch
these options at run-time.