[LibTracyClient] Enable SW Timer fallback (#6664)
This includes a change to the selected feature flags: the SW timer
fallback is enabled by adding `TRACY_TIMER_FALLBACK`.

This fallback is most likely to kick in on VM's that may not have
support for the `rdtsc` instruction. The fallback SW timer can lead
to a dramatic drop in resolution (8 nanoseconds with rdtsc -> 15.6
*milliseconds* on the machine tested), but thankfully the Tracy Profiler
GUI is pretty good about making the timer resolution error bars
noticeable and encoding a resolution measurement into the trace.

Any user that hits this fallback is likely to notice the loss of fidelity
and still get _some_ useful information out of their trace, so I think
it's a good option to automatically switch to the fallback.

In any case, the "error reporting" we get from an self-initializing
C library is of course to crash the whole application, so even without
the fallback we would have had to find a better way to communicate the
error regarding the missing `rdtsc` support.