[Bison] avoid adding `Libiconv` as a dependency on Linux
This is an attempt to avoid errors such as the following:
```
julia> using Bison_jll
       run(`$(Bison_jll.bison()) --version`)
Inconsistency detected by ld.so: dl-version.c: 204: _dl_check_map_versions: Assertion `needed != NULL' failed!
```

My hope is that this error is triggered by the fact that glibc contains
its own iconv implementation, and the fact that we're linking in
`libiconv` in addition to glibc causes something to break.  I have other
binaries that are built against `Libiconv_jll` and work just fine so I'm
not sure why this one in particular breaks.