TracyProfiler: Use rr-safe nopl; rdtsc sequence (#7225)
Several users have reported masssive slowdowns to rr record/replay
in packages that have the tracy client loaded. The issue turns out
to be an unlucky false positive in an rr heuristic that determines
the patchability of `rdtsc`. In rr master, it is possible to guarantee
patchability by using a specific `nopl; rdtsc` sequence. Add a patch
to tracy to do just that.

See also
https://github.com/rr-debugger/rr/pull/3580
https://github.com/JuliaLang/julia/pull/50975