CUDA: Make use of a local toolkit require a separate preference.
That way, the 'version' preference is always set to a valid version.
This makes it so that CUDA.jl will know the version of the local toolkit.

To make this feature more powerful, add some libcudart-based detection
of the local toolkit's version. This should make setting `local = "true"`
behave similar to using artifacts, i.e., if the driver/runtime are
available during precompilation, the version will be auto-detected.
Only for environments without CUDA (such as containers, log-in nodes, etc)
the user will need to set the `version` preference to inform the system
what version of CUDA is being used.

This PR also adds / changes artifact comparison strategies to always
bail out when using a local toolkit. This should prevent downloading
useless artifacts when using system libraries.

[skip build]