[libsingular_julia] update to latest, drop Singular_jll compat entry
Decoupling libsingular_julia_jll from Singular_jll is OK because
Singular.jl also specifies a version of Singular_jll. The drawback
is that one could end up accidentally using the wrong Singular_jll
version. But the big advantage is that we are freed to work with
newer Singular_jll release in Singular.jl which is safe when using
the JLL override mechanism there (when changes in the C++ code is
detected while in dev mode, the C++ code just gets compiled and
used to replace the code in the JLL).