[BOLT] Add builder (#8391)
* Add BOLT to LLVM_full

 Tested locally with x86_64-linux-gnu-cxx11.

* Build BOLT on all platforms

In the discord, I see someone built on FreeBSD and seems to build on windows here: https://github.com/llvm/llvm-project/issues/73085.
I haven't tested locally, as this takes hours to build for me.
Might need a patch for windows: https://github.com/llvm/llvm-project/issues/73006.

* Fix typo

* Build Bolt on >= LLVM 14

Not sure why the code in common.jl uses `-gt` so much.

* Don't build on LLVM 14.0.5

Seems like you cannot build on it and it shouldn't even exist in Yggdrasil.

* Only build on LLVM >= 16

* Patch for ambiguity in `timespec` on FreeBSD builder

* Backport fix to windows install

* Disable building runtime library on windows

Presumably this also prevents it being built on FreeBSD but according to https://reviews.llvm.org/D148742 it's not supported anyways.

* Revert "Patch for ambiguity in `timespec` on FreeBSD builder"

FreeBSD fails even with this patch and we've disabled building runtime as it's not supported.
So this patch is useless.
This reverts commit 9aa1a0771757d3ccbd74ee266b1652d8bbc813f6.

* Fix windows install patch

* Trigger remaining builds

* Only build BOLT on supported arches and OS

Not sure if there's a better approach. I don't know if calling build_tarballs twice will cause problems.

* Fix build script for assert build

* Fix typo

* Build BOLT separately from LLVM

* Disable i686 build

Cross compilation for it seems to be broken, https://github.com/llvm/llvm-project/issues/84547.

* Only build and install bolt binaries + libraries

Cuts tarball size from ~400 mb to 32mb. Untarred llvm-bolt and llvm-bolt-heatmap are both 42mb and 99% of size.

* Enable building bolt runtime and clean up script

* Set Julia Compat to v1.6

Co-authored-by: Mosè Giordano <giordano@users.noreply.github.com>

* Add all the binaries to the products

Might as well, won't be hard to adjust if they are removed at some point.

* Clarify comment

* Format

* Bump to 18.1.4

---------

Co-authored-by: Mosè Giordano <giordano@users.noreply.github.com>