Add builder for the zstd seekable format subproject
This is an extension of the zstd format and lives in zstd's `contrib/`
directory. It's not built by default and has no build system of its own
nor integration with the parent build system, so here it's built with a
single manual call to `cc`.

There doesn't appear to be a single, official name for this subproject.
I've opted to name the library `zstd_seekable` in keeping with its
associated header file, `zstd_seekable.h`, and the prefix used for the
library's symbols (`ZSTD_seekable` for all but a couple).

This could be built and distributed by Zstd_jll but I made this
standalone since the seekable format is not technically part of zstd
proper (though it would be more convenient).