[LibSSH2] Re-build to lock compatibility with OpenSSL v3.0.14+
In https://github.com/JuliaPackaging/Yggdrasil/pull/8393 we changed the layout of OpenSSL to put libraries in `${prefix}/lib` rather than `${prefix}/lib64`.  While this doesn't effect Julia much, if one uses a tool such a `JLLPrefixes` to plop a bunch of JLL's artifacts in a single prefix and then try to use the binaries directly, the RPATHs no longer work.  This rebuild will create a new version of LibSSH2 that will work with the new OpenSSL binaries.