Upgrade to Wayland v1.23
At the same time, resolve a bootstrap issue wherein we cannot ask for
the correct version of the host-system `wayland-scanner` if the host
packages have not yet been built. Instead, perform a host-system
compilation of just a utility `wayland-scanner`, and then in a second
step perform the cross-compile of all of the Wayland libraries and a
target-compatible `wayland-scanner`.

- Note that the bootstrapping `wayland-scanner` purposely disables the
  DTD validation feature in order to work around a problem with properly
  linking the host's libxml2 when cross-compiling for FreeBSD.