[terminfo-db] Add a builder for the terminfo database
This uses the directory-based database rather than the single Berkeley
database (terminfo.db). The license is from ncurses, which is close
enough; see https://invisible-island.net/ncurses/terminfo.src.html.

Ideally this would be shipped as one of Julia's dependencies to ensure
that the custom terminfo parser actually works on all platforms, as not
all have the terminfo database.