Add a builder for the AWS command line interface v2
This is based in part on
https://github.com/aws/aws-cli/discussions/7168, which is for FreeBSD
but should apply to any other platform as well (or so I would assume).
It differs from the process described in the discussion in that it uses
a provided installation script rather than the development build
instructions in the README, which is all the process in the discussion
is, except with some particular requisite system package installs.
Notable among the installed packages are CMake and Rust, which is why
I've added Rust to the `compilers` list and defined some CMake
environment variables.