[XZ] Use default GCC version instead of v7
The `preferred_gcc_version` was set to v7 in PR 8166 due to issues with
the assembler emitting instructions not supported by the version of
binutils bundled with the default GCC. This was in service of updating
XZ to v5.6.0, but as of PR 8396, we build XZ v5.2.5. When v2.5.2 was
initially added in PR 2057, we used the default GCC, so it should no
longer be necessary to require v7.