Pick up another rr commit (#9478)
The way our CI system uses --pack-dir (which was a patch we had been carrying until it got merged upstream) does not quite match what ended up upstream. I'm submitting a PR upstream to fix that, but in the meantime, this carries the changes to make our existing CI system work as before.