Replace use of environment variables with build-time settings
This avoids the need for an `__init__` method in the JLL. This is also
similar to the method already (implicitly) used by the OpenCL loader.
Instead, use a combination of settings that still install headers and
libraries into the target build prefix while also specifying the
destination users' system configuration and data directories (where
loader descriptions will be found).