[Xorg_libxcb] Rebuild for freebsd-aarch64 (#9870)
In checking that the dependencies were properly built, it seems that
the dependecy on XSLT_jll is unnecessary. The in-source `ChangeLog`
file (which has more detail than the git-repo NEWS [1]) says that
XSLT has been made optional since v1.10 in 2013, and the Arch Linux
package [2] agrees in making it a build-only dependecy. See also
[3] where it is mentioned that it's only used for generating HTML
checks results.

[1] https://gitlab.freedesktop.org/xorg/lib/libxcb/-/blob/master/NEWS
[2] https://archlinux.org/packages/extra/x86_64/libxcb/
[3] https://bugs.freedesktop.org/show_bug.cgi?id=23863