[GAP_pkg_*] change GAP_lib_jll to a BuildDependency
Logically the code in these packages only depends on the GAP
kernel ABI, and hence GAP_jll, but not on the library code in
GAP_lib_jll. Decoupling these make it easier to perform various
upgrades.

This changes has no effect yet, but the next time we update any
of the GAP_pkg_* JLLs we will see the benefit.