[OCaml] Add fallback for re-located libdir / bindir
The existing `OCAMLLIB` environment variable already allows you to use
a standard library installed elsewhere, but it doesn't let you re-locate
an existing installation. As one might expect, the problem is that the
stdlib itself contains absolute references to its own paths.

This patch tries to hook into both places where that happens and do an
on-the-fly replacement of the standard library path, if it turns out
this is the standard library that OCaml that was built with (just
moved).

This kicks in if the default hard-coded `%%LIBDIR%%`/`%%BINDIR%%` cannot
be found.