[GoogleCodeSearch] update to force Go toolchain update
The go version and stdlib this was compiled with 5 years ago has several reported vulnerabilties.
Since JuliaHub uses this jll, we get pinged on this in our docker image scans all the time.

I'm not sure of the best way to force a rebuild, but it doesn't build as it was. 
I have to set julia_compat to 1.6 to have it build for the experimental platforms.

Is this the best way to do force a rebuild? Will this cause a released jll as `1.2.0+1` when it's merged?