bootstrap: Build libstdc++ to be compatible with both TLS and non-TLS ABI
As reported in https://github.com/JuliaLang/julia/issues/57021, our Yggdrasil-built
binaries no longer work out-of-the box with the msys2-provided mingw toolchain.
This is not a good situation, because it basically means that partial
BinaryBuilder-provided builds are off the table. The root cause of the issue
is that at some point msys2 started building gcc (and libstdc++) with
TLS support. Enabling TLS is an ABI-breaking change, because it removes
symbols from libstdc++ that were referenced in the header. That said,
I don't believe it is necessary for this to be ABI breaking. In particular,
I think the included patch provides a correct ABI-compability implementation
that supports applications linked against either the TLS enabled or disabled
version. The net effect of this change would be that (after this is bumped
on CSL) the new binaries we build would have TLS enabled and work on
the new msys64 ABI as well as the new CSL, which any old binaries could
continue to work on either the new CSL or the old CSL.
Notably, new binaries would not work on the old CSL, so once we bump
CSL, we'll have to put in new JLL compats for any newly built JLLs.

Note that this only changes the libstdc++ for GCC@14, which is the one
we bundle in CSL. However, it does change the TLS flag for all versions
of GCC, so that TLS will become enabled even when the new binaries
are built with older GCC.