0_RootFS: Update MinGW sources
As noted in https://github.com/JuliaLang/julia/issues/56840, our CSL-distributed msvcrt
is no longer compatible with mingw as distributed in msys2. This is because
https://github.com/mingw-w64/mingw-w64/commit/91b83c26858857c3af0ac285c30c2a785e420358
added a call to _initterm_e but the corresponding msvcrt.a emulation was added
in https://github.com/mingw-w64/mingw-w64/commit/6c89679ccc51ea4f218683e61943881a7038a1c9
which is newer than what we ship in CSL.

That said, it is not clear to me why we ship msvcrt.a from mingw in CSL at all. It makes
sense to ship gnd dlls, but if we're not shipping a toolchain it's odd to ship msvcrt.a.
After all, we're not shipping glibc either.