Add recipe for clang-extract (#12165)
This adds a recipe for https://github.com/SUSE/clang-extract. The repo comes with a meson build script, but I wasn't able to get it to work, because of both JuliaPackaging/BinaryBuilderBase.jl#445 and because it's trying to look for llvm-config of the llvm install, which we don't ship for cross-compile reasons. Instead, I had Claude just write a simple build script that compiles everything together manually. There's only a handful of C++ files, so I think this is simplest.