[ROCm] introduce HSARuntime_jll
My eventual goal here is to ship the entire ROCm stack repackaged from
ROCm nightlies. The HSA runtime is needed to determine the ROCm platform
we should ship artifacts for. The next step would be to introduce a
`rocm_sdk_core_jll` that uses platform augmentation based on this jll to
select the right artifact, which I already have a proof of concept for
locally.